move MathSelector and MathPropertySelector.js into own file
