+++ frontend/src/scenes/insights/ActionFilter/MathSelector.js
@@ -1,0 +1,82 @@
+import React from 'react'
+import { Tooltip, Select } from 'antd'
+import { PROPERTY_MATH_TYPE, EVENT_MATH_TYPE, MATHS } from 'lib/constants'
+
+const EVENT_MATH_ENTRIES = Object.entries(MATHS).filter(([, item]) => item.type == EVENT_MATH_TYPE)
+const PROPERTY_MATH_ENTRIES = Object.entries(MATHS).filter(([, item]) => item.type == PROPERTY_MATH_TYPE)
+
+export function MathSelector({ math, index, onMathSelect, areEventPropertiesNumericalAvailable, style }) {
+    const numericalNotice = `This can only be used on properties that have at least one number type occurence in your events.${
+        areEventPropertiesNumericalAvailable ? '' : ' None have been found yet!'
+    }`
+
+    return (
+        <Select
+            style={{ width: 150, ...style }}
+            value={math || 'total'}
+            onChange={(value) => onMathSelect(index, value)}
+            data-attr={`math-selector-${index}`}
+        >
+            <Select.OptGroup key="event aggregates" label="Event aggregation">
+                {EVENT_MATH_ENTRIES.map(([key, { name, description, onProperty }]) => {
+                    const disabled = onProperty && !areEventPropertiesNumericalAvailable
+                    return (
+                        <Select.Option key={key} value={key} data-attr={`math-${key}-${index}`} disabled={disabled}>
+                            <Tooltip
+                                title={
+                                    onProperty ? (
+                                        <>
+                                            {description}
+                                            <br />
+                                            {numericalNotice}
+                                        </>
+                                    ) : (
+                                        description
+                                    )
+                                }
+                                placement="right"
+                            >
+                                <div
+                                    style={{
+                                        height: '100%',
+                                        width: '100%',
+                                        paddingRight: 8,
+                                        overflow: 'hidden',
+                                        textOverflow: 'ellipsis',
+                                    }}
+                                >
+                                    {name}
+                                </div>
+                            </Tooltip>
+                        </Select.Option>
+                    )
+                })}
+            </Select.OptGroup>
+            <Select.OptGroup key="property aggregates" label="Property aggregation">
+                {PROPERTY_MATH_ENTRIES.map(([key, { name, description, onProperty }]) => {
+                    const disabled = onProperty && !areEventPropertiesNumericalAvailable
+                    return (
+                        <Select.Option key={key} value={key} data-attr={`math-${key}-${index}`} disabled={disabled}>
+                            <Tooltip
+                                title={
+                                    onProperty ? (
+                                        <>
+                                            {description}
+                                            <br />
+                                            {numericalNotice}
+                                        </>
+                                    ) : (
+                                        description
+                                    )
+                                }
+                                placement="right"
+                            >
+                                <div style={{ height: '100%', width: '100%' }}>{name}</div>
+                            </Tooltip>
+                        </Select.Option>
+                    )
+                })}
+            </Select.OptGroup>
+        </Select>
+    )
+}
